fix data types for sqlc generation

+++ db/migration/000001_init_schema.up.sql
@@ -1,21 +1,21 @@
 CREATE TABLE "accounts" (
     "id" bigserial PRIMARY KEY,
     "owner" varchar NOT NULL,
-    "balance" decimal NOT NULL,
+    "balance" bigint NOT NULL,
     "currency" varchar NOT NULL,
     "created_at" timestamptz NOT NULL DEFAULT (now())
 );
 CREATE TABLE "entries" (
     "id" bigserial PRIMARY KEY,
     "account_id" bigint NOT NULL,
-    "amount" decimal NOT NULL,
+    "amount" bigint NOT NULL,
     "created_at" timestamptz NOT NULL DEFAULT (now())
 );
 CREATE TABLE "transfers" (
     "id" bigserial PRIMARY KEY,
     "from_account_id" bigint NOT NULL,
     "to_account_id" bigint NOT NULL,
-    "amount" decimal NOT NULL,
+    "amount" bigint NOT NULL,
     "created_at" timestamptz NOT NULL DEFAULT (now())
 );
 CREATE INDEX ON "accounts" ("owner");
